Make a huge mess out of 'oc' dependencies
This drug a bunch of stuff in with it because I used
ParseAndConvertConfig from controller/common in the machine-config-operator.
If I port just that function in (or find something in ignition itself I
could use), I think I can drop some of these dependencies.

This was a shiftweek fun project, I'll clean it up later. :)